Partly moved to new more return semantics
Simpler approach... none of this abort calling mode nonsense,
instead modes just return a simple value which is for use
of that mode and the caller (e.g. which player scored in a game.

Not necessary to return anythin from run...

Still need to change over ContinuousMode and PongMatch modes